remove paintcache
