+++ Mapsui.Rendering.Skia/SkiaStyles/LineStringRenderer.cs
@@ -29,7 +29,7 @@ public static class LineStringRenderer
         var lineWidth = (float)(vectorStyle.Line?.Width ?? 1f);
         if (vectorStyle.Line.IsVisible())
         {
-            using var paint = renderService.PaintCache.GetOrCreatePaint((vectorStyle.Line, opacity), CreateSkPaint);
+            using var paint = renderService.VectorCache.GetOrCreatePaint((vectorStyle.Line, opacity), CreateSkPaint);
             using var path = renderService.VectorCache.GetOrCreatePath((feature.Id, position, extent, rotation, lineWidth), ToPath);
             canvas.DrawPath(path, paint);
         }
