+++ Mapsui.Rendering.Skia/SkiaStyles/SymbolStyleRenderer.cs
@@ -147,13 +147,13 @@ public class SymbolStyleRenderer : ISkiaStyleRenderer, IFeatureSize
         using var path = vectorCache.GetOrCreatePath(symbolStyle.SymbolType, CreatePath);
         if (symbolStyle.Fill.IsVisible())
         {
-            using var fillPaint = vectorCache.PaintCache.GetOrCreatePaint((symbolStyle.Fill!, opacity), CreateFillPaint);
+            using var fillPaint = vectorCache.GetOrCreatePaint((symbolStyle.Fill!, opacity), CreateFillPaint);
             canvas.DrawPath(path, fillPaint);
         }
 
         if (symbolStyle.Outline.IsVisible())
         {
-            using var linePaint = vectorCache.PaintCache.GetOrCreatePaint((symbolStyle.Outline!, opacity), CreateLinePaint);
+            using var linePaint = vectorCache.GetOrCreatePaint((symbolStyle.Outline!, opacity), CreateLinePaint);
             canvas.DrawPath(path, linePaint);
         }
